Refactor token handling and improve error responses in user authentication

+++ src/utils/cloudinary.js
@@ -18,7 +18,9 @@ import fs from 'fs';
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if(localFilePath!=null){
+        fs.unlinkSync(localFilePath) 
+        }// remove the locally saved temporary file as the upload operation got failed
         return null;
     }
 }
